show flair for each issue
comments and github reactions, as varying-size arc segments around the
issue

this makes it easy to see external interest in issues. initial positive
feedback: it showed the 'docker compose support for task definitions',
which is clearly a highly desired feature, but not yet mapped out. prior
to this change it was lost among the sea of issues.

next steps:

* fix bounding box so that the flair doesn't get cut off
* sort graphs by size + flair, with flair as the tiebreaker?