generalize intermediate graph construction
the number of edges adjacent to a node has impact on its construction.
previously, we would construct an intermediate graph of Cards, and then
map those Cards into actual Nodes. this makes it hard to have other
types in the intermediate form, so instead construct a graph of thunks
which are later applied with their context to produce the
already-general Node type.

also, store all of 'data' on the model, rather than plucking off fields.
this makes it more obvious which parts are "interpreted" and which are
straight from the backend.