refresh when hook is received
this is kind of painful since it's effectively a bridge between the v3
REST API and the v4 GraphQL API, but hopefully worth it

it's also kind of coarse at the moment; when any card is dragged it'll
fetch all of them. i'm gonna push this for now though and see that
things at least work before optimizing.