implement card accept/rejecting
accepting adds an 'accepted' label and closes the issue

rejecting adds a 'rejected' label, moves it back to the backlog (if
there's only one)

the rejected label sticks around until it's then accepted. i thought
about having it go away when the issue is re-started, but i'm curious to
see if this way is more informational/useful (e.g. as a 'warning' to the
pm to carefully accept it)