dim filtered-out cards in graphs
this is a little bit janky-janky, but it's better than nothing. it's
nice that it keeps context for things like the project graph page, but
boy do we have some large graphs.

also, made the main graph view take up the full page, just like the
projects page. limiting it to 980 looks a little silly on a large
display.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>