fix 'this week' logic for archive/dashboard
archive just took past 7 days, dashboard just took events until it saw a
Sunday, both are subtly wrong.

instead, reverse the order of the archive so it's easier to use
'takeWhile' and change eventsThisWeek to use proper time logic (thanks
to Time.Extra), and use it for both the archive and dashboard.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>