collect rotation history of card assignees
not currently used for anything yet, but this data can be used to keep
track of how often people pair with each other, which could be used to
influence automatic rotation/pair shuffling

it also includes the timing information (i.e. how long the rotation
lasted), which i don't really have a use case for right now, but the
data could be interesting in the future. (time has to be taken into
account regardless as a way to correlate assignments/unassignments
together as one "rotation," so we might as well just keep the data.)

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>