include cards in 'done' column in Pairs page
issues are only truly done when they're closed, but can still be 'done'
in the sense that a completed PR has been submitted and the pair has
moved on to the next issue.

we can keep track of this intermediate state as an open issue in the
'done' column, which will be closed upon accepting the PR (or manually).
until then, users may be assigned to this already-'done' issue to
represent going through the PR review (which should now also show up in
the card itself) or other acceptance criteria.

also: tweak spacing

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>