implement support for template projects
issues can now have a project created for them based on a pre-existing
template. projects named 'Template: Foo' will be selectable on a 'new
project' button on issue cards within the same repo. the new project
will be named 'Foo #(issue number)' and will be displayed on the issue
card.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>