add apply/cancel interaction
show shuffled pairs immediately, and save changes on apply. this way you
can just keep quickly shuffling and then finally save at the end.

note: the 'pretend' arrangement stays after hitting apply - you have to
cancel to reset it. this feels better than clearing it out and letting
it trickle-update, which can make it hard to follow whether it stuck
with the shuffled order or not.

after clicking cancel, sometimes the order of the cards changes, since
the moves were sent in parallel and don't specify order. let's call this
a quirk, not a bug. :)

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>